refactor: reorganize mcp module structure
