+++ python/mcp_server/__init__.py
@@ -14,6 +14,7 @@
 
 from .server import MCPComponentServer
 from .client import MCPClient
+from .ai_search.ai_search_server import AIsearch
 
 
-__all__ = ["MCPComponentServer", "MCPClient"]
+__all__ = ["MCPComponentServer", "MCPClient", "AIsearch"]
