+++ python/mcp_server/ai_search/__init__.py
@@ -1,0 +1,9 @@
+"""
+AI search module for MCP server.
+
+This module provides functionality for searching and retrieving AI-related data and models.
+"""
+
+from .ai_search_server import AIsearch, server
+
+__all__ = ["AIsearch", "server"]
