+++ python/tests/test_ai_search_mcp_stdio.py
@@ -37,8 +37,8 @@ class TestGetAppList(unittest.TestCase):
             assert result.content[0] != ""
 
         subprocess.check_call([sys.executable, "-m", "pip", "install", "mcp"])
-        from appbuilder.mcp import MCPClient
-        from appbuilder.mcp import ai_search_server
+        from appbuilder.mcp_server import MCPClient
+        from appbuilder.mcp_server.ai_search import ai_search_server
 
         loop = asyncio.get_event_loop()
         loop.run_until_complete(process())
