+++ python/tests/test_appbuilder_client_mcp.py
@@ -87,7 +87,7 @@ class TestAppBuilderClientMCP(unittest.TestCase):
             await appbuilder_client.http_client.session.close()
             await mcp_client.cleanup()
 
-        from appbuilder.mcp.client import MCPClient
+        from appbuilder.mcp_server.client import MCPClient
         subprocess.check_call([sys.executable, "-m", "pip", "install", "mcp"])
         loop = asyncio.get_event_loop()
         loop.run_until_complete(process())
