+++ python/tests/test_appbuilder_client_mcp_component.py
@@ -90,7 +90,7 @@ class TestAgentRuntime(unittest.TestCase):
             [sys.executable, "-m", "pip", "uninstall", "-y", "chainlit"]
         )
         subprocess.check_call([sys.executable, "-m", "pip", "install", "mcp"])
-        from appbuilder.mcp.client import MCPClient
+        from appbuilder.mcp_server.client import MCPClient
         asyncio.run(handler())
